+++ batchglm/train/tf/base_glm_all/__init__.py
@@ -1,4 +1,4 @@
 from .estimator import EstimatorAll
 from .estimator_graph import EstimatorGraphAll
 from .jacobians import JacobiansTF
-from .hessians import HessianTF
+from .hessians import HessianGLMALL
